[FE-217] Refactoring usersFinancicalTransactions, Main and UserActions.
