--- src/actions/userBorrowOperations.js
@@ -1,59 +1,0 @@
-import { web3FromAddress } from '@polkadot/extension-dapp';
-import {
-	BORROW_REQUEST_START,
-	BORROW_REQUEST_ERROR,
-	BORROW_REQUEST_SUCCESS,
-} from './types';
-import API from '../services';
-
-export function borrow(account, keyring, underlyingAssetId, borrowAmount) {
-	return async (dispatch) => {
-		const callBack = ({ events = [], status }) => {
-			if (status.isFinalized) {
-				events.forEach(
-					({
-						event: {
-							method,
-							section,
-							data: [error],
-						},
-					}) => {
-						if (section === 'system' && method === 'ExtrinsicSuccess') {
-							dispatch({
-								type: BORROW_REQUEST_SUCCESS,
-							});
-						} else if (method === 'ExtrinsicFailed' && error.isModule) {
-							const decoded = API.registry.findMetaError(error.asModule);
-							const { documentation } = decoded;
-							dispatch({
-								type: BORROW_REQUEST_ERROR,
-								payload: documentation.join(' '),
-							});
-						}
-					}
-				);
-			}
-		};
-
-		try {
-			dispatch({ type: BORROW_REQUEST_START });
-			const currentUser = keyring.getPair(account);
-
-			if (currentUser.isLocked) {
-				const injector = await web3FromAddress(account);
-				await API.tx.minterestProtocol
-					.borrow(underlyingAssetId, borrowAmount)
-					.signAndSend(account, { signer: injector.signer }, callBack);
-			} else {
-				await API.tx.minterestProtocol
-					.borrow(underlyingAssetId, borrowAmount)
-					.signAndSend(currentUser, callBack);
-			}
-		} catch (err) {
-			dispatch({
-				type: BORROW_REQUEST_ERROR,
-				payload: err.toString(),
-			});
-		}
-	};
-}
