+++ src/components/UserActions/BorrowOperations/BorrowOperations.jsx
@@ -4,11 +4,11 @@ import SendBorrow from '../../Forms/SendBorrow/SendBorrow';
 import classes from './BorrowOperations.module.css';
 
 export default function BorrowOperations(props) {
-	const { account, keyring, borrow, isBorrowResponseRunning } = props;
+	const { keyring, account, borrow, isBorrowResponseRunning } = props;
 
 	const handleSendBorrow = (form) => {
 		const { underlyingAssetId, borrowAmount } = form;
-		borrow(account, keyring, underlyingAssetId, borrowAmount);
+		borrow(keyring, account, underlyingAssetId, borrowAmount);
 	};
 	return (
 		<div className={classes.deposit}>
