+++ src/components/UserActions/Redeem/Redeem.jsx
@@ -4,11 +4,11 @@ import SendRedeem from '../../Forms/SendRedeem/SendRedeem';
 import classes from './Redeem.module.css';
 
 export default function BorrowOperations(props) {
-	const { account, keyring, redeem, isRedeemResponseRunning } = props;
+	const { keyring, account, redeem, isRedeemResponseRunning } = props;
 
 	const handleSendRedeem = (form) => {
 		const { underlyingAssetId } = form;
-		redeem(account, keyring, underlyingAssetId);
+		redeem(keyring, account, underlyingAssetId);
 	};
 	return (
 		<div className={classes.redeem}>
