+++ src/components/UserActions/Repay/Repay.jsx
@@ -4,11 +4,11 @@ import SendRepay from '../../Forms/SendRepay/SendRepay';
 import classes from './Repay.module.css';
 
 export default function Repay(props) {
-	const { account, keyring, repay, isRepayResponseRunning } = props;
+	const { keyring, account, repay, isRepayResponseRunning } = props;
 
 	const handleSendRepay = (form) => {
 		const { underlyingAssetId, repayAmount } = form;
-		repay(account, keyring, underlyingAssetId, repayAmount);
+		repay(keyring, account, underlyingAssetId, repayAmount);
 	};
 	return (
 		<div className={classes.repay}>
