+++ src/components/UserActions/RepayAll/RepayAll.jsx
@@ -4,11 +4,11 @@ import SendRepayAll from '../../Forms/SendRepayAll/SendRepayAll';
 import classes from './RepayAll.module.css';
 
 export default function RepayAll(props) {
-	const { account, keyring, repayAll, isRepayAllResponseRunning } = props;
+	const { keyring, account, repayAll, isRepayAllResponseRunning } = props;
 
 	const handleSendRepayAll = (form) => {
 		const { underlyingAssetId } = form;
-		repayAll(account, keyring, underlyingAssetId);
+		repayAll(keyring, account, underlyingAssetId);
 	};
 	return (
 		<div className={classes.repay}>
