+++ src/redux/index.js
@@ -6,10 +6,7 @@ import accountReducer from './reducers/accountReducer';
 import substrateReducer from './reducers/substrateReducer';
 import economicUpdatesReducer from './reducers/economicUpdatesReducer';
 import adminReducer from './reducers/adminReducer';
-import userDepositOperationsReducer from './reducers/userDepositOperationsReducer';
-import userRedeemOperationsReducer from './reducers/userRedeemOperationsReducer';
-import userBorrowOperationsReducer from './reducers/userBorrowOperationsReducer';
-import userRepayOperationsReducer from './reducers/userRepayOperationsReducer';
+import usersFinancicalTransactionsReducer from './reducers/usersFinancicalTransactionsReducer';
 import dashboardDataReducer from './reducers/dashboardDataReducer';
 
 const reducers = {
@@ -18,10 +15,7 @@ const reducers = {
 	substrate: substrateReducer,
 	economicUpdates: economicUpdatesReducer,
 	admin: adminReducer,
-	userDepositOperations: userDepositOperationsReducer,
-	userRedeemOperations: userRedeemOperationsReducer,
-	userBorrowOperations: userBorrowOperationsReducer,
-	userRepayOperations: userRepayOperationsReducer,
+	usersFinancicalTransactions: usersFinancicalTransactionsReducer,
 	dashboardData: dashboardDataReducer,
 };
